+++ source/Messaging.Infrastructure/Configuration/InternalCommands/CommandExecutor.cs
@@ -15,6 +15,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
+using Messaging.Application.Common.Commands;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Messaging.Infrastructure.Configuration.InternalCommands
