+++ source/Messaging.Infrastructure/Configuration/InternalCommands/CommandScheduler.cs
@@ -14,6 +14,7 @@
 
 using System;
 using System.Threading.Tasks;
+using Messaging.Application.Common.Commands;
 using Messaging.Infrastructure.Configuration.DataAccess;
 using Messaging.Infrastructure.Configuration.Serialization;
 using NodaTime;
