+++ source/Messaging.IntegrationTests/Infrastructure/InternalCommands/InternalCommandProcessorTests.cs
@@ -15,6 +15,7 @@
 using System.Data;
 using System.Threading.Tasks;
 using Dapper;
+using Messaging.Application.Common.Commands;
 using Messaging.Application.Configuration.DataAccess;
 using Messaging.Infrastructure.Configuration.InternalCommands;
 using Messaging.IntegrationTests.Fixtures;
